fix: converted rolldown config to mjs bc of type issue with premove
